+++ packages/framework/lib/components/error_manager/error_manager.dart
@@ -5,7 +5,7 @@ export 'catcher_error_manager.dart';
 abstract class ErrorManager {
   ErrorManager();
 
-  Future<void> init(VoidCallback runAppFunction);
+  Future<void> init(AsyncCallback runAppFunction);
 
   void reportCheckedError(dynamic error, dynamic stackTrace);
 }