+++ packages/flrx_test/CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.2
+Fixed: TestErrorManager init signature
 
-* TODO: Describe initial release.
+## 0.0.1
+initial release.